Fix _iterencode_once: move nonlocal to separate module for Python 2
Python 2 raises SyntaxError on nonlocal even inside an if-guarded
block, since the parser processes the entire function body.  Move the
cycle breaker wrapper to simplejson/_cycle_breaker.py which is only
imported on Python 3.
