+++ simplejson/encoder.py
@@ -47,6 +47,15 @@ try:
 except NameError:
     _dict_types = (dict,)
 
+# Break the reference cycle between mutually recursive closures in
+# _make_iterencode after encoding completes (CPython 3.15 json module).
+# _wrap_iterencode_once uses nonlocal (Python 3-only syntax), so it
+# lives in a separate module to avoid SyntaxError on Python 2.
+if PY3:
+    from ._cycle_breaker import _wrap_iterencode_once
+else:
+    _wrap_iterencode_once = None
+
 def encode_basestring(s, _PY3=PY3, _q=u'"'):
     """Return a JSON representation of a Python string
 
@@ -768,18 +777,7 @@ def _make_iterencode(markers, _default, _encoder, _indent, _floatstr,
                     if markers is not None:
                         del markers[markerid]
 
-    if _PY3:
-        def _iterencode_once(o, _current_indent_level):
-            """Wrapper that breaks the reference cycle between the mutually
-            recursive closures after the top-level call completes.  Without
-            this, the closures prevent prompt garbage collection."""
-            nonlocal _iterencode, _iterencode_dict, _iterencode_list
-            try:
-                for chunk in _iterencode(o, _current_indent_level):
-                    yield chunk
-            finally:
-                del _iterencode, _iterencode_dict, _iterencode_list
-
-        return _iterencode_once
-
+    if _wrap_iterencode_once is not None:
+        return _wrap_iterencode_once(
+            _iterencode, _iterencode_dict, _iterencode_list)
     return _iterencode
